Delete some old 4DN code...
Add `getControllers` method to DefaultItemView that allows
usage of reusable state logic via controller components.
Hopefully should allow simpler transition to functional component
views in future (if doing that).
