+++ src/encoded/static/components/static-pages/HomePage/index.js
@@ -54,7 +54,8 @@ const HomePageTitle = React.memo(function HomePageTitle(props){
     return (
         <PageTitleContainer alerts={alerts}>
             <TitleAndSubtitleUnder subtitle="Clinical Genomics Analysis Platform" className="home-page-title">
-                <strong>TODO:</strong> Portal Title Here
+                Chromosaur
+                <sup className="small text-200"> (working name)</sup>
             </TitleAndSubtitleUnder>
         </PageTitleContainer>
     );
